Create README.md
I have a partner and a soulmate. I need descretness and mutual respect honesty and I'll give the same in return. I can anything you want me to be and I won't do anything the you don't want to have done . Imm 40 honest and want to help give and be me . I'm kind gentle and have a great personality easy on eyes and on decent shape 6 foot 3 inches  and 8 between my legs.